Refactor and centralize plotting in MULTINEST workflow
Moved spectrum, chemistry, surface-albedo, and posterior plotting from __multinest.py to shared helpers in __plotting.py, reducing code duplication and improving maintainability. Updated all references to use the new centralized plotting functions. Updated version to 2.5.5 and revised documentation and changelog accordingly.
