+++ README.md
@@ -2,7 +2,7 @@
 
 # ExoReL<sup>R</sup>
 
-Version 2.5.4
+Version 2.5.5
 
 Includes:
 * A routine to generate of reflected light spectra;
