+++ __init__.py
@@ -1,3 +1,3 @@
-__version__ = '2.5.4'
+__version__ = '2.5.5'
 
 from .__main__ import *
